Move Action type to index file
